requirements.txt
flask>=3.0.0
gunicorn>=21.2.0
anthropic>=0.40.0
requests>=2.31.0
python-dotenv>=1.0.0
msal>=1.24.0
python-docx>=1.1.0
markdown>=3.5.0
reportlab>=4.1.0
psycopg2-binary>=2.9.9
